llm-d-inference-sim-custom-dataset/pyproject.toml
[project]
name = "llm-d-inference-sim-custom-dataset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.43",
]
